PSMS/BE/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pymysql
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
